# app/Http/Controllers/AuthController.php
<?php

namespace App\Http\Controllers;

use App\Http\Requests\LoginRequest;
use App\Http\Requests\RegisterRequest;
use App\Models\User;
use Illuminate\Http\Request;
use Illuminate\Support\Facades\Hash;
use Illuminate\Validation\ValidationException;

class AuthController extends Controller
{
    /**
     * @OA\Post(
     *     path="/api/register",
     *     summary="Регистрация нового пользователя",
     *     tags={"Auth"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"name", "email", "password", "password_confirmation"},
     *             @OA\Property(property="name", type="string", example="Анна Петрова"),
     *             @OA\Property(property="email", type="string", format="email", example="anna@example.com"),
     *             @OA\Property(property="password", type="string", example="password123"),
     *             @OA\Property(property="password_confirmation", type="string", example="password123"),
     *             @OA\Property(property="phone", type="string", example="+79991234567")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Успешная регистрация",
     *         @OA\JsonContent(
     *             @OA\Property(property="message", type="string", example="User registered successfully.")
     *         )
     *     )
     * )
     */
    public function register(RegisterRequest $request)
    {
        $user = User::create([
            'name' => $request->name,
            'phone' => $request->phone,
            'email' => $request->email,
            'password' => bcrypt($request->password),
            'role' => 'user',
        ]);

        auth()->login($user);

        $token = $user->createToken('api-token')->plainTextToken;

        return response()->json([
            'message' => 'User registered and logged in.',
            'token'   => $token,
        ], 201);
    }

    /**
     * @OA\Post(
     *     path="/api/login",
     *     summary="Авторизация пользователя",
     *     tags={"Auth"},
     *     @OA\RequestBody(
     *         required=true,
     *         @OA\JsonContent(
     *             required={"email", "password"},
     *             @OA\Property(property="email", type="string", format="email", example="user@example.com"),
     *             @OA\Property(property="password", type="string", example="password123"),
     *             @OA\Property(property="session-token", type="string", example="abc123xyz")
     *         )
     *     ),
     *     @OA\Response(
     *         response=200,
     *         description="Токен авторизации",
     *         @OA\JsonContent(
     *             @OA\Property(property="token", type="string", example="1|XoG...token...")
     *         )
     *     ),
     *     @OA\Response(
     *         response=422,
     *         description="Неверные учетные данные"
     *     )
     * )
     */
    public function login(LoginRequest $request)
    {
        $user = User::where('email', $request->email)->first();

        if (! $user || ! Hash::check($request->password, $user->password)) {
            throw ValidationException::withMessages([
                'email' => ['Invalid credentials.'],
            ]);
        }

        // 💡 Логика слияния корзины
        if ($request->has('session-token')) {
            $guestCart = \App\Models\Cart::where('session_token', $request->get('session-token'))->first();
            if ($guestCart && !$guestCart->user_id) {
                $existingUserCart = \App\Models\Cart::where('user_id', $user->id)->first();

                if ($existingUserCart) {
                    // Переносим товары из гостевой корзины в корзину пользователя
                    foreach ($guestCart->items as $item) {
                        $existingUserCart->items()->updateOrCreate(
                            ['product_id' => $item->product_id],
                            ['quantity' => $item->quantity]
                        );
                    }
                    $guestCart->delete(); // Удаляем гостевую корзину
                } else {
                    // Просто привязываем гостевую корзину к пользователю
                    $guestCart->update([
                        'user_id' => $user->id,
                        'session_token' => null,
                    ]);
                }
            }
        }

        // создается API-токен
        return response()->json([
            'token' => $user->createToken('api-token')->plainTextToken,
        ]);
    }


    /**
     * @OA\Get(
     *     path="/api/me",
     *     summary="Получить текущего пользователя",
     *     tags={"Auth"},
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Текущий пользователь",
     *         @OA\JsonContent(ref="#/components/schemas/User")
     *     )
     * )
     */
    public function me(Request $request)
    {
        return response()->json($request->user());
    }

    /**
     * @OA\Post(
     *     path="/api/logout",
     *     summary="Выход из системы (удаляет все токены)",
     *     tags={"Auth"},
     *     security={{"sanctum":{}}},
     *     @OA\Response(
     *         response=200,
     *         description="Выход выполнен успешно"
     *     )
     * )
     */
    public function logout(Request $request)
    {
        $request->user()->tokens()->delete();

        return response()->json(['message' => 'Logged out']);
    }
}